File re-arrange
