+++ 32blit/32blit.hpp
@@ -1,10 +1,9 @@
 #pragma once
 
 #include "engine/engine.hpp"
-#include "engine/utility.hpp"
 #include "engine/output.hpp"
 #include "engine/input.hpp"
-#include "engine/audio.hpp"
+#include "audio/audio.hpp"
 #include "engine/timer.hpp"
 #include "engine/tweening.hpp"
 #include "graphics/blend.hpp"
