+++ 32blit/engine/output.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../types/pixel_format.hpp"
+#include "../graphics/surface.hpp"
 #include <stdint.h>
 
 namespace blit {
