+++ 32blit/graphics/color.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../types/pixel_format.hpp"
+#include "surface.hpp"
 
 namespace blit {
