+++ 32blit/math/geometry.cpp
