+++ 32blit/math/geometry.hpp
